requirements.txt
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.2
mysql-connector==2.2.9
mysql-connector-python==8.0.31
numpy==1.24.1
packaging==22.0
Pillow==9.3.0
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
